Fix duplicate linkage nodes
Previously, we picked the most general unary_compare overload that
compared the addess of Linkage to Linkage::rep. As a result, those never
compared equal and we ended up creating a new node for linkage each time
Lexicon::cxx_linkage was called. This patch adds an overload to resolve
this problem.